requirements.txt
asyncio
yt-dlp
python-dotenv
spotipy
discord
discord.py
PyNaCl